pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "interview-scheduler"
version = "0.1.0"
description = "CLI tool for scheduling interviews with Google Calendar integration"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "interview_scheduler"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
click = ">=8.0.0,<8.2.0"
rich = "^13.0.0"
jinja2 = "^3.1.0"
google-api-python-client = "^2.0.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.0.0"
pyyaml = "^6.0"
appdirs = "^1.4.4"
python-dateutil = "^2.8.0"
pytz = "^2023.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
interview-scheduler = "interview_scheduler.cli:app"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=interview_scheduler --cov-report=term-missing"
